d_a_npc_uk: _create 97.15->97.76 (getType u8->int - retail switches on unmasked r3; remaining = shared-error-tail placement); .o 02:05

+++ include/d/actor/d_a_npc_uk.h
@@ -99,7 +99,7 @@ public:
     void setEyePos(cXyz i_eyePos) { mEyePos = i_eyePos; }
     void setFlag(u16 flag) { mFlags |= flag; }
 
-    u8 getType();
+    int getType();
     u8 getPath();
     int getShapeType();
     u16 getCaughtFlag();
